city-guides/requirements.txt
Flask>=2.3.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
